MON-3802: implement cross-namespace rules for UWM
This change introduces a way to deploy user-defined rules which are not
scoped to their namespace of origin.

To enable the feature, a user-defined monitoring admin needs to
configure at least one namespace in the UWM ConfigMap:

```
kind: ConfigMap
apiVersion: v1
metadata:
  name: user-workload-monitoring-config
  namespace: openshift-user-workload-monitoring
data:
  config.yaml: |-
    namespacesWithoutLabelEnforcement: [ user-monitoring-shared ]
```

For all `PrometheusRule` objects defined in the `user-monitoring-shared`
namespace, Prometheus and Thanos Ruler evaluate the PromQL expressions
without enforcing the namespace label of origin. It makes it possible to
have generic rules that get applied to all (or a subset of) the user
projects instead of having individual rule objects in each user project.
The rules can also query for platform namespaces only.

The capability is enabled by default but a cluster admin can decide to
disable it in the CMO ConfigMap:

```
kind: ConfigMap
apiVersion: v1
metadata:
  name: cluster-monitoring-config
  namespace: openshift-monitoring
data:
  config.yaml: |-
    userWorkloadEnabled: true
    userWorkload:
     rulesWithoutLabelEnforcementAllowed: false
```

For example, a user-defined admin can create a single rule that fires
when a user namespace doesn't enforce the Restricted pod security
policy.

```
apiVersion: monitoring.coreos.com/v1
kind: PrometheusRule
metadata:
  name: security
  namespace: user-monitoring-shared
spec:
  groups:
    - name: pod-security-policy
      rules:
        - alert: "NamespaceNotEnforcingRestrictedPolicy"
          expr: kube_namespace_labels{namespace!~"(openshift|kube).*|default",label_pod_security_kubernetes_io_enforce!="restricted"}
          for: 5m
          annotations:
            summary: "Restricted policy not enforced"
            description: "Namespace {{ $labels.namespace }} doesn't enforce the Restricted pod security policy."
          labels:
            severity: warning
```

Signed-off-by: Simon Pasquier <spasquie@redhat.com>